+++ CHANGELOG.md
@@ -56,6 +56,7 @@ Master list of UniV3 forks:
   * Add `DODOV1` action to Polygon
   * Add `DODOV1` action to Scroll
 * Add `rebateClaimer()(address)` function on Mainnet Settlers for gas rebate program
+* Add SolidlyV3 UniV3 fork to Sonic
 
 ## 2024-12-14